requirements.txt
typer
rich
google-generativeai
python-dotenv
importlib-metadata
